third_party/blink/web_tests/platform/generic/external/wpt/css/css-shapes/shape-outside/values/shape-margin-002-expected.txt

This is a testharness.js-based test.
FAIL 10.12345px - inline assert_equals: expected "10.12345px" but got "10.1235px"
PASS .5px - inline
PASS +15px - inline
PASS +10.678px - inline
FAIL 10.12345px - computed assert_equals: expected "10.123px" but got "10.124px"
PASS .5px - computed
PASS +15px - computed
PASS +10.678px - computed
Harness: the test ran to completion.
